Moved answers summary to partial template
This lays the groundwork for a separate 'saved answers' page
