+++ app/controllers/oidc_controller.rb
@@ -4,6 +4,7 @@ class OidcController < ApplicationController
   def callback
     reset_session
     sub = request.env['omniauth.auth']['extra']['raw_info']['sub']
+    session[:nhs_sub] = sub
     saved_answer = FormResponse.where(NHSSub: sub).first
     if saved_answer
       session[:saved_answer] = true
@@ -11,15 +12,15 @@ class OidcController < ApplicationController
       redirect_to check_your_answers_url
     else
       session[:saved_answer] = false
+
+      session[:auto_generated_data] = {
+        example_1: 'Lorem ipsum est...',
+        example_2: 'Something or other'
+      }
+      session[:userinfo] = request.env['omniauth.auth'].info
+
       redirect_to address_url
     end
-
-    session[:auto_generated_data] = {
-      example_1: 'Lorem ipsum est...',
-      example_2: 'Something or other'
-    }
-    session[:nhs_sub] =
-      session[:userinfo] = request.env['omniauth.auth'].info
   end
 
   def failure; end
