+++ config/locales/en.yml
@@ -30,28 +30,32 @@
 # available at https://guides.rubyonrails.org/i18n.html.
 
 en:
-  check_your_answers:
+  answers_summary:
+    auto_generated_info_heading: Information generated by this application
+    nhs_user_info_heading: Information provided from your nhs account
+    form_info_heading: Information filled in on this form
+    auto_generated_data:
+      example_1:
+        label: 'Auto generated example'
+      example_2:
+        label: 'Another auto generated example'
     userinfo:
       email:
         label: Email
       phone_number:
         label: Phone number
+    help_with_food_delivery:
+      required:
+        label: 'Help with food delivery required'
+  check_your_answers:
     title: Are you ready to send your application?
     description: Check your answers first.
-    nhs_user_info_heading: Information provided from your nhs account
-    auto_generated_info_heading: Information generated by this application
-    form_info_heading: Information filled in on this form
     send_heading: Now send your application
     confirmation: By submitting this application you’re confirming that, as far as you know, the details you’re providing are correct.
     submit: Accept and send
     help_with_food_delivery:
       required:
         label: 'Help with food delivery required'
-    auto_generated_data:
-      example_1:
-        label: 'Auto generated example'
-      example_2:
-        label: 'Another auto generated example'
   questions:
     address:
       title: What is your address?
